refactor: consolidate slide elements into shared module
